refactor(#5): split reference detail sheet into story/project/files cards
Keep the dialog as orchestrator; extract the three content cards for smaller reviews.
